Uses registers API to get data about job families and roles
This large commit adds code to access the registers api, retrieve data, and return it back to the user. It removes the need to manually update a database with content. However, it's likely that since other professions are not as advanced, there will still need to be a database of some sort for them. As it stands, this is a good starting point.